+++ R/ContextualImportanceUtility.R
@@ -75,9 +75,7 @@
 #' @references Främling, K. *Contextual Importance and Utility in R: the 'ciu' Package.*
 #' In: Proceedings of 1st Workshop on Explainable Agency in Artificial Intelligence,
 #' at 35th AAAI Conference on Artificial Intelligence. Virtual, Online. February 8-9, 2021. pp. 110-114.
-#' @references Främling, K. *Explainable AI without Interpretable Model*. 2020, <https://arxiv.org/abs/2009.13996>.
-#' @references Främling, K. *Decision Theory Meets Explainable AI*. 2020, <doi.org/10.1007/978-3-030-51924-7_4>.
-#' @references Främling, K. *Modélisation et apprentissage des préférences par réseaux de neurones pour l'aide à la décision multicritère*. 1996, <https://tel.archives-ouvertes.fr/tel-00825854/document> (title translation in English: *Learning and Explaining Preferences with Neural Networks for Multiple Criteria Decision Making*)
+#' @references Främling, K. *Decision Theory Meets Explainable AI*. 2020, <doi:10.1007/978-3-030-51924-7_4>.
 #' @examples
 #' # Explaining the classification of an Iris instance with lda model.
 #' # We use a versicolor (instance 100).
@@ -212,12 +210,14 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
     else if ( inherits(o.model, "lm") ) { # lm
       o.predict.function <- function(model, inputs) { predict(model, inputs) }
     }
-    else {
-      # This works at least with "lda" model, don't know with which other ones.
+    else if ( inherits(o.model, "lda") ) { # lm
       o.predict.function <- function(model, inputs) {
         pred <- predict(model,inputs)
         return(pred$posterior)
       }
+    }
+    else {
+      o.predict.function <- predict
     }
   }
   else {
@@ -469,7 +469,7 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
     rows <- sample(nrow(reps))
     reps <- reps[rows,]
 
-    # Have to restore "ordered" columsn if there are any
+    # Have to restore "ordered" columns if there are any
     for ( i in 1:length(ind.inputs) )
       if ( is.ordered(instance[,ind.inputs[i]]) )
         reps[,ind.inputs[i]] <- as.ordered(reps[,ind.inputs[i]])
@@ -565,7 +565,9 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
   # Set neutral.CU="" to avoid having a neutral, orange line.
   ggplot.ciu <- function(instance, ind.input=1, ind.output=1, in.min.max.limits=NULL,
                          n.points=40, main=NULL, xlab="x", ylab="y", ylim=NULL,
-                         illustrate.CIU=FALSE, neutral.CU=0.5, CIU.illustration.colours=c("red", "orange", "green", "blue")) {
+                         illustrate.CIU=FALSE, neutral.CU=0.5,
+                         CIU.illustration.colours=c("red", "orange", "green", "blue"),
+                         categorical_style="segment") {
     # Bogus line just to get rid of strange NOTE i Check: "Undefined global functions or variables: x y"
     x <- y <- 0
 
@@ -578,15 +580,16 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
         instance[,ind.input] <- factor(instance[,ind.input], o.inp.levels[[ind.input]])
     }
 
+    # Check and set up minimum/maximum limits for inputs
+    if ( is.null(in.min.max.limits) )
+      in.min.max.limits <- o.in.minmax[ind.input,]
+    if ( is.null(in.min.max.limits) )
+      stop("No minimum/maximum limits provided to 'new' nor 'plot.ciu'")
+    in.min <- in.min.max.limits[1]
+    in.max <- in.min.max.limits[2]
+
     # First deal with "numeric" possibility.
     if ( is.numeric(instance[,ind.input]) ) {
-      # Check and set up minimum/maximum limits for inputs
-      if ( is.null(in.min.max.limits) )
-        in.min.max.limits <- o.in.minmax[ind.input,]
-      if ( is.null(in.min.max.limits) )
-        stop("No minimum/maximum limits provided to 'new' nor 'plot.ciu'")
-      in.min <- in.min.max.limits[1]
-      in.max <- in.min.max.limits[2]
       interv <- (in.max - in.min)/n.points
       xp <- seq(in.min,in.max,interv)
     }
@@ -640,15 +643,32 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
     # Create plot, show current value
     df <- data.frame(x=xp, y=yp[,ind.output])
     cdf <- data.frame(x=instance[,ind.input], y=as.numeric(cu.val[ind.output]))
-    p <- ggplot(df, aes(x=x, y=y)) +
-      labs(title=main, x=xlab, y=ylab)
     if ( is.numeric(instance[,ind.input]) ) {
-      p <- p + geom_line()
+      p <- ggplot(df, aes(x=x, y=y)) + geom_line() + ylim(ylim)
     }
     else { # factor
-      p <- p + geom_col()
+      if ( !is.null(categorical_style) && categorical_style == "segment" ) {
+        p <- ggplot(df, aes(x=as.numeric(x), y=y))
+        p <- p + geom_segment(aes(x = as.numeric(x) - 0.5, xend = as.numeric(x) + 0.5,
+                                  y = y, yend = y)) +  # Centered horizontal segments
+          # geom_segment(aes(x = as.numeric(x) + 0.5, xend = as.numeric(x) + 0.5,
+          #                  y = y, yend = lead(y, 1, default = tail(y, 1)))) +  # Vertical segments between steps
+          geom_segment(aes(x = as.numeric(x) + 0.5, xend = as.numeric(x) + 0.5,
+                           y = y, yend = c(y[-1], utils::tail(y, 1)))) +
+          scale_x_continuous(breaks = 1:length(df$x), labels = df$x) + # Custom X-axis labels
+          ylim(ylim)
+      }
+      else {
+        # This doesn't work anymore, for some reason. Or it works but somehow the "ylim" breaks it.
+        # Still OK to do ylim on the return ggplot. Strange!!!???
+        p <- ggplot(df, aes(x=x, y=y)) +
+          geom_col() #+
+        #scale_y_continuous() +
+        #ylim(ylim)
+      }
     }
-    p <- p + geom_point(data=cdf, colour = "red", size=4) + lims(y = ylim)
+    p <- p + geom_point(data=cdf, colour = "red", size=4) +
+      labs(title=main, x=xlab, y=ylab)
 
     # Illustrate CIU calculation?
     if ( illustrate.CIU ) {
@@ -672,6 +692,8 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
       if ( cmax - cdf$y > cdf$y - cmin ) vjust <- "bottom" else vjust <- "top"
       p <- p + annotate("text", x=in.max, y=cdf$y, label="y", colour=CIU.illustration.colours[4],
                         vjust = vjust, hjust = "inward", fontface="italic")
+
+      # is.numeric() mainly checks if neutral.cu==NULL or similar.
       if ( is.numeric(neutral.CU)) {
         neutral <- cmin + neutral.CU*(cmax - cmin)
         if ( cmax - neutral > neutral - cmin ) vjust <- "bottom" else vjust <- "top"
@@ -990,10 +1012,9 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
         main=main.title, ...)
   }
 
-  #' Create `ciu` object from this `CIU` object.
-  #'
-  #' @return `ciu` object
-  #' @export
+  # Create `ciu` object from this `CIU` object.
+  #
+  # @return `ciu` object
   as.ciu <- function() {
     ciu <- list(
       model = o.model,
@@ -1036,8 +1057,11 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
       plot.ciu(instance, ind.input, ind.output, in.min.max.limits, n.points, main, xlab, ylab, ylim, ...)
     },
     ggplot.ciu = function(instance, ind.input=1, ind.output=1, in.min.max.limits=NULL, n.points=40, main=NULL, xlab=NULL, ylab=NULL,
-                          ylim=NULL, illustrate.CIU=FALSE, neutral.CU=0.5, CIU.illustration.colours=c("red", "orange", "green", "blue")) {
-      ggplot.ciu(instance, ind.input, ind.output, in.min.max.limits, n.points, main, xlab, ylab, ylim, illustrate.CIU, neutral.CU, CIU.illustration.colours)
+                          ylim=NULL, illustrate.CIU=FALSE, neutral.CU=0.5, CIU.illustration.colours=c("red", "orange", "green", "blue"),
+                          categorical_style="segment") {
+      ggplot.ciu(instance, ind.input, ind.output, in.min.max.limits, n.points, main,
+                 xlab, ylab, ylim, illustrate.CIU, neutral.CU, CIU.illustration.colours,
+                 categorical_style)
     },
     plot.ciu.3D = function(instance, ind.inputs, ind.output, in.min.max.limits=NULL, n.points=40,
                            main=NULL, xlab=NULL, ylab=NULL, zlab=NULL, zlim=NULL, ...) {
@@ -1075,7 +1099,7 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
                               cu.colours = c("darkgreen", "darkgreen", "0.8"),
                               low.color="red", mid.color="yellow",
                               high.color="darkgreen",
-                              use.influence=FALSE,
+                              use.influence=FALSE, scale.CI=FALSE,
                               sort=NULL, decreasing=FALSE, # These are not used yet.
                               main=NULL) {
       ciu.ggplot.col(as.ciu(), instance, ind.inputs, output.names, in.min.max.limits,
@@ -1083,7 +1107,7 @@ ciu.new <- function(bb, formula=NULL, data=NULL, in.min.max.limits=NULL, abs.min
                      show.input.values, concepts.to.explain,
                      target.concept, target.ciu, ciu.meta, plot.mode, ci.colours, cu.colours,
                      low.color, mid.color, high.color,
-                     use.influence,
+                     use.influence, scale.CI,
                      sort, decreasing, main)
     },
     textual = function(instance=NULL, ind.inputs=NULL, ind.output=1,